#71 Move useful API from draftcxx
moved to spec model bundle
